Working

+++ preperation for final/lab04/p06/VecInt.hpp
@@ -1,7 +1,7 @@
 #ifndef VECINT_H
 #define VECINT_H
-#include <stddef.h>
 
+#include <stddef.h>
 struct VecInt
 {
     int *data;
@@ -10,7 +10,7 @@ struct VecInt
 };
 
 void VecInt_createEmpty(struct VecInt *self);
-void VecInt_createOfSize(struct VecInt *self, int size, int initValue);
+void VecInt_createOfSize(struct VecInt *self, size_t size, int initValue);
 void VecInt_pushBack(struct VecInt *self, int x);
 void VecInt_destroy(struct VecInt *self);
